Row-bind a list of data.frames or remote tables

Row-bind a list of data.frames or remote tables

Usage

bind_source(data, ..., force_proceed = getOption("odcfun.force_proceed"))

Arguments

data

A list of data.frame or remote tables, e.g., output from execute_def().

...

Named arguments for each variable included in the output. The argument name should be the new name in the output, and the right hand side of the argument is a character vector of the original names. The name vector and the list elements in data will be matched by position. if an output variable only came from some of the sources, fill the name vector to a length equal to the number of sources with NA, e.g., var only come from the second out of three sources, then var = c(NA, 'nm_in_src2', NA).

force_proceed

A logical for whether to ask for user input in order to proceed when remote tables are needed to be collected for binding. The default is TRUE to let user be aware of that the downloading process may be slow. Use options(odcfun.force_proceed = FALSE) to suppress the prompt once and for all.

Value

A data.frame or remote table containing combined rows of the input list with variables specified by ...

Examples

df1 <- subset(iris, Species == "setosa")
df2 <- subset(iris, Species == "versicolor")
df3 <- subset(iris, Species == "virginica")

bind_source(list(df1, df2, df3),
  s_l = "Sepal.Length",
  s_w = "Sepal.Width",
  p_l_setosa = c("Petal.Length", NA, NA),
  p_l_virginica = c(NA, NA, "Petal.Length")
)
#>     src_No s_l s_w p_l_setosa p_l_virginica
#> 1        1 5.1 3.5        1.4            NA
#> 2        1 4.9 3.0        1.4            NA
#> 3        1 4.7 3.2        1.3            NA
#> 4        1 4.6 3.1        1.5            NA
#> 5        1 5.0 3.6        1.4            NA
#> 6        1 5.4 3.9        1.7            NA
#> 7        1 4.6 3.4        1.4            NA
#> 8        1 5.0 3.4        1.5            NA
#> 9        1 4.4 2.9        1.4            NA
#> 10       1 4.9 3.1        1.5            NA
#> 11       1 5.4 3.7        1.5            NA
#> 12       1 4.8 3.4        1.6            NA
#> 13       1 4.8 3.0        1.4            NA
#> 14       1 4.3 3.0        1.1            NA
#> 15       1 5.8 4.0        1.2            NA
#> 16       1 5.7 4.4        1.5            NA
#> 17       1 5.4 3.9        1.3            NA
#> 18       1 5.7 3.8        1.7            NA
#> 19       1 5.1 3.8        1.5            NA
#> 20       1 5.4 3.4        1.7            NA
#> 21       1 5.1 3.7        1.5            NA
#> 22       1 4.6 3.6        1.0            NA
#> 23       1 5.1 3.3        1.7            NA
#> 24       1 4.8 3.4        1.9            NA
#> 25       1 5.0 3.0        1.6            NA
#> 26       1 5.0 3.4        1.6            NA
#> 27       1 5.2 3.5        1.5            NA
#> 28       1 5.2 3.4        1.4            NA
#> 29       1 4.7 3.2        1.6            NA
#> 30       1 4.8 3.1        1.6            NA
#> 31       1 5.4 3.4        1.5            NA
#> 32       1 5.2 4.1        1.5            NA
#> 33       1 5.5 4.2        1.4            NA
#> 34       1 5.0 3.2        1.2            NA
#> 35       1 5.5 3.5        1.3            NA
#> 36       1 4.9 3.6        1.4            NA
#> 37       1 4.4 3.0        1.3            NA
#> 38       1 5.1 3.4        1.5            NA
#> 39       1 5.0 3.5        1.3            NA
#> 40       1 4.5 2.3        1.3            NA
#> 41       1 4.4 3.2        1.3            NA
#> 42       1 5.0 3.5        1.6            NA
#> 43       1 5.1 3.8        1.9            NA
#> 44       1 5.1 3.8        1.6            NA
#> 45       1 4.6 3.2        1.4            NA
#> 46       1 5.3 3.7        1.5            NA
#> 47       1 5.0 3.3        1.4            NA
#> 48       2 7.0 3.2         NA            NA
#> 49       2 6.4 3.2         NA            NA
#> 50       2 6.9 3.1         NA            NA
#> 51       2 5.5 2.3         NA            NA
#> 52       2 6.5 2.8         NA            NA
#> 53       2 5.7 2.8         NA            NA
#> 54       2 6.3 3.3         NA            NA
#> 55       2 4.9 2.4         NA            NA
#> 56       2 6.6 2.9         NA            NA
#> 57       2 5.2 2.7         NA            NA
#> 58       2 5.0 2.0         NA            NA
#> 59       2 5.9 3.0         NA            NA
#> 60       2 6.0 2.2         NA            NA
#> 61       2 6.1 2.9         NA            NA
#> 62       2 5.6 2.9         NA            NA
#> 63       2 6.7 3.1         NA            NA
#> 64       2 5.6 3.0         NA            NA
#> 65       2 5.8 2.7         NA            NA
#> 66       2 6.2 2.2         NA            NA
#> 67       2 5.6 2.5         NA            NA
#> 68       2 5.9 3.2         NA            NA
#> 69       2 6.1 2.8         NA            NA
#> 70       2 6.3 2.5         NA            NA
#> 71       2 6.4 2.9         NA            NA
#> 72       2 6.6 3.0         NA            NA
#> 73       2 6.8 2.8         NA            NA
#> 74       2 6.7 3.0         NA            NA
#> 75       2 6.0 2.9         NA            NA
#> 76       2 5.7 2.6         NA            NA
#> 77       2 5.5 2.4         NA            NA
#> 78       2 6.0 2.7         NA            NA
#> 79       2 5.4 3.0         NA            NA
#> 80       2 6.0 3.4         NA            NA
#> 81       2 6.3 2.3         NA            NA
#> 82       2 5.5 2.5         NA            NA
#> 83       2 5.5 2.6         NA            NA
#> 84       2 6.1 3.0         NA            NA
#> 85       2 5.8 2.6         NA            NA
#> 86       2 5.0 2.3         NA            NA
#> 87       2 5.6 2.7         NA            NA
#> 88       2 5.7 3.0         NA            NA
#> 89       2 5.7 2.9         NA            NA
#> 90       2 6.2 2.9         NA            NA
#> 91       2 5.1 2.5         NA            NA
#> 92       3 6.3 3.3         NA           6.0
#> 93       3 5.8 2.7         NA           5.1
#> 94       3 7.1 3.0         NA           5.9
#> 95       3 6.3 2.9         NA           5.6
#> 96       3 6.5 3.0         NA           5.8
#> 97       3 7.6 3.0         NA           6.6
#> 98       3 4.9 2.5         NA           4.5
#> 99       3 7.3 2.9         NA           6.3
#> 100      3 6.7 2.5         NA           5.8
#> 101      3 7.2 3.6         NA           6.1
#> 102      3 6.5 3.2         NA           5.1
#> 103      3 6.4 2.7         NA           5.3
#> 104      3 6.8 3.0         NA           5.5
#> 105      3 5.7 2.5         NA           5.0
#> 106      3 5.8 2.8         NA           5.1
#> 107      3 6.4 3.2         NA           5.3
#> 108      3 6.5 3.0         NA           5.5
#> 109      3 7.7 3.8         NA           6.7
#> 110      3 7.7 2.6         NA           6.9
#> 111      3 6.0 2.2         NA           5.0
#> 112      3 6.9 3.2         NA           5.7
#> 113      3 5.6 2.8         NA           4.9
#> 114      3 7.7 2.8         NA           6.7
#> 115      3 6.3 2.7         NA           4.9
#> 116      3 6.7 3.3         NA           5.7
#> 117      3 7.2 3.2         NA           6.0
#> 118      3 6.2 2.8         NA           4.8
#> 119      3 6.1 3.0         NA           4.9
#> 120      3 6.4 2.8         NA           5.6
#> 121      3 7.2 3.0         NA           5.8
#> 122      3 7.4 2.8         NA           6.1
#> 123      3 7.9 3.8         NA           6.4
#> 124      3 6.3 2.8         NA           5.1
#> 125      3 6.1 2.6         NA           5.6
#> 126      3 7.7 3.0         NA           6.1
#> 127      3 6.3 3.4         NA           5.6
#> 128      3 6.4 3.1         NA           5.5
#> 129      3 6.0 3.0         NA           4.8
#> 130      3 6.9 3.1         NA           5.4
#> 131      3 6.7 3.1         NA           5.6
#> 132      3 6.9 3.1         NA           5.1
#> 133      3 6.8 3.2         NA           5.9
#> 134      3 6.7 3.0         NA           5.2
#> 135      3 6.3 2.5         NA           5.0
#> 136      3 6.5 3.0         NA           5.2
#> 137      3 6.2 3.4         NA           5.4
#> 138      3 5.9 3.0         NA           5.1