Grouping records into episodes by date range

This function is useful for collapsing, e.g., medication dispensation or hospitalization, records into episodes if the records' dates are no more than n days gap apart. The length of the gap can be relaxed by another grouping variable.

Usage

collapse_episode(
  data,
  clnt_id,
  start_dt,
  end_dt = NULL,
  gap,
  overwrite = NULL,
  gap_overwrite = Inf,
  .dt_trans = data.table::as.IDate,
  ...
)

Arguments

data

A data.frame that contains the id and date variables.

clnt_id

Column name of subject/person ID.

start_dt

Column name of the starting date of records.

end_dt

Column name of the end date of records. The default is NULL assuming the record last one day and only the start date will be used to calculate the gaps between records.

gap

A number in days that will be used to separate episodes. It should be the inclusive upper bound + 1. For example, gap = 8 means collapsing records no more than 7 days apart.

overwrite

Column name of a grouping variable determining whether the consecutive records are related and should have a different gap value. For example, dispensing records may have the same original prescription number, and a different gap value can be assigned for situation, e.g., the days between two records is > gap, but these records still belong to the same prescription.

gap_overwrite

A different gap value used for related records. The default is Inf, which means all records with the same overwrite variable will be collapsed.

.dt_trans

Function to transform start_dt/end_dt. Default is [data.table::as.IDate()].

...

Additional arguments passing to the .dt_trans function.

Value

The original data.frame with new columns indicating episode grouping

Examples

# make toy data
df <- make_test_dat()

head(df)
#>   uid clnt_id      dates diagx diagx_1 diagx_2
#> 1  20       1 2018-09-14  3042    3057     999
#> 2   6       3 2016-05-13  3056    3043     999
#> 3  95       3 2016-08-29   999     999     999
#> 4  33       4 2018-01-21  3049    3056    <NA>
#> 5  60       4 2018-11-17   999    <NA>    <NA>
#> 6  51       4 2019-04-28   999    <NA>    <NA>

# collapse records no more than 90 days apart
collapse_episode(df, clnt_id, dates, gap = 91)
#>     uid clnt_id      dates diagx diagx_1 diagx_2 epi_id epi_no epi_seq
#> 1    20       1 2018-09-14  3042    3057     999      1      1       1
#> 2     6       3 2016-05-13  3056    3043     999      2      1       1
#> 3    95       3 2016-08-29   999     999     999      3      2       1
#> 4    33       4 2018-01-21  3049    3056    <NA>      4      1       1
#> 5    60       4 2018-11-17   999    <NA>    <NA>      5      2       1
#> 6    51       4 2019-04-28   999    <NA>    <NA>      6      3       1
#> 7    30       5 2016-10-23  3044    3056     999      7      1       1
#> 8    87       5 2017-08-04   999     999     999      8      2       1
#> 9    99       5 2017-09-03   999    <NA>    <NA>      8      2       2
#> 10   61       5 2018-08-06   999    <NA>    <NA>      9      3       1
#> 11   92       5 2019-09-10   999     999    <NA>     10      4       1
#> 12    7       5 2020-01-01  3049    3050     999     11      5       1
#> 13   15       6 2019-01-06   305    3041    <NA>     12      1       1
#> 14   46       7 2015-05-07  3046    3049    3044     13      1       1
#> 15   16       7 2015-12-19  3040    3046    <NA>     14      2       1
#> 16    4       7 2016-11-02  3046    3056    <NA>     15      3       1
#> 17   65       9 2020-08-11   999    <NA>     999     16      1       1
#> 18   98      10 2015-08-30   999    <NA>    <NA>     17      1       1
#> 19   10      10 2016-05-13   304    3050    <NA>     18      2       1
#> 20   13      10 2016-05-27  3042    3040    <NA>     18      2       2
#> 21   72      10 2016-12-25   999     999    <NA>     19      3       1
#> 22   62      12 2019-10-31   999     999    <NA>     20      1       1
#> 23   83      13 2019-10-30   999     999     999     21      1       1
#> 24   97      13 2020-07-12   999     999    <NA>     22      2       1
#> 25   93      14 2016-02-02   999    <NA>     999     23      1       1
#> 26   29      15 2016-01-18  3044    3046    <NA>     24      1       1
#> 27   48      16 2017-03-16  3047    3053    3042     25      1       1
#> 28   37      16 2018-01-18  3041    3054     999     26      2       1
#> 29   64      17 2018-05-03   999    <NA>    <NA>     27      1       1
#> 30   12      17 2020-04-09  3054    3048     999     28      2       1
#> 31  100      19 2016-11-28   999    <NA>     999     29      1       1
#> 32   53      19 2016-12-06   999     999    <NA>     29      1       2
#> 33   39      20 2019-12-23  3043    3057    <NA>     30      1       1
#> 34   71      22 2015-10-29   999     999    <NA>     31      1       1
#> 35   77      22 2017-08-27   999     999    <NA>     32      2       1
#> 36   57      22 2019-05-01   999     999    <NA>     33      3       1
#> 37   59      23 2015-02-02   999     999    <NA>     34      1       1
#> 38   96      23 2016-12-14   999     999    <NA>     35      2       1
#> 39   14      23 2019-12-18  3045    3053     999     36      3       1
#> 40   21      23 2020-04-11  3046    3054    <NA>     37      4       1
#> 41   25      24 2016-11-16  3043    3045    <NA>     38      1       1
#> 42   68      24 2018-05-17   999    <NA>     999     39      2       1
#> 43   27      25 2015-08-13  3046    3058     999     40      1       1
#> 44   94      26 2016-07-01   999    <NA>    <NA>     41      1       1
#> 45   90      26 2018-07-20   999    <NA>     999     42      2       1
#> 46   50      27 2016-03-25  3041    3047    3040     43      1       1
#> 47   43      27 2019-09-24  3054    3044     305     44      2       1
#> 48    5      28 2015-07-16  3045     999    <NA>     45      1       1
#> 49    8      29 2015-02-11   305    <NA>     999     46      1       1
#> 50   22      30 2017-04-17  3051    3048     999     47      1       1
#> 51    1      31 2017-01-23  3058     304    <NA>     48      1       1
#> 52   91      31 2019-06-24   999    <NA>     999     49      2       1
#> 53   45      32 2015-02-28  3052    3042    3042     50      1       1
#> 54   31      33 2016-03-16  3054     305     999     51      1       1
#> 55   70      33 2016-06-05   999    <NA>     999     51      1       2
#> 56   54      34 2015-11-02   999     999    <NA>     52      1       1
#> 57   47      34 2017-02-07  3055    3041    3043     53      2       1
#> 58   69      34 2019-01-08   999     999    <NA>     54      3       1
#> 59   28      35 2019-05-07  3055    3052     999     55      1       1
#> 60   55      36 2015-09-04   999     999    <NA>     56      1       1
#> 61   82      36 2016-03-03   999     999    <NA>     57      2       1
#> 62   26      36 2018-11-02  3058    3047    <NA>     58      3       1
#> 63   18      36 2019-08-24  3050    3053     999     59      4       1
#> 64   42      37 2016-01-21  3049    3041    3050     60      1       1
#> 65   56      38 2016-01-24   999     999    <NA>     61      1       1
#> 66    9      38 2016-07-12  3054    3049     999     62      2       1
#> 67   88      39 2015-10-16   999    <NA>     999     63      1       1
#> 68   23      39 2016-10-22   305    3048     999     64      2       1
#> 69   76      39 2016-11-04   999     999    <NA>     64      2       2
#> 70   24      39 2017-08-07  3053    3047    <NA>     65      3       1
#> 71   52      39 2019-09-24   999     999     999     66      4       1
#> 72   89      39 2020-09-16   999    <NA>     999     67      5       1
#> 73   86      39 2020-12-09   999    <NA>    <NA>     67      5       2
#> 74   79      40 2016-08-27   999     999     999     68      1       1
#> 75   67      41 2015-10-04   999    <NA>    <NA>     69      1       1
#> 76   38      41 2019-04-12  3043    3046    <NA>     70      2       1
#> 77   17      42 2016-06-02  3052    3055     999     71      1       1
#> 78   81      42 2016-11-17   999     999    <NA>     72      2       1
#> 79    3      42 2017-04-18  3052    3044    <NA>     73      3       1
#> 80   74      42 2018-10-01   999    <NA>     999     74      4       1
#> 81   19      43 2017-02-18  3041    3051    <NA>     75      1       1
#> 82    2      43 2017-10-03  3053     305    <NA>     76      2       1
#> 83   41      43 2017-12-13  3058    3045    3043     76      2       2
#> 84   85      43 2020-07-03   999    <NA>     999     77      3       1
#> 85   58      44 2015-11-21   999     999     999     78      1       1
#> 86   73      44 2017-06-21   999    <NA>     999     79      2       1
#> 87   35      44 2017-11-08   304    3053     999     80      3       1
#> 88   40      45 2016-12-29  3055     999    <NA>     81      1       1
#> 89   63      46 2015-04-28   999     999    <NA>     82      1       1
#> 90   11      46 2016-01-10  3054    3046     999     83      2       1
#> 91   75      46 2017-10-26   999     999    <NA>     84      3       1
#> 92   49      46 2017-11-08  3055    3059    3051     84      3       2
#> 93   80      47 2016-08-08   999    <NA>    <NA>     85      1       1
#> 94   36      47 2016-08-21  3043    3058    <NA>     85      1       2
#> 95   78      47 2016-10-24   999     999     999     85      1       3
#> 96   66      47 2020-07-14   999     999    <NA>     86      2       1
#> 97   44      47 2020-10-23  3043    3054    3040     87      3       1
#> 98   34      48 2016-08-06  3053    3056    <NA>     88      1       1
#> 99   32      49 2019-08-23  3058    3049    <NA>     89      1       1
#> 100  84      50 2015-07-12   999     999     999     90      1       1
#>     epi_start_dt epi_stop_dt
#> 1     2018-09-14  2018-09-14
#> 2     2016-05-13  2016-05-13
#> 3     2016-08-29  2016-08-29
#> 4     2018-01-21  2018-01-21
#> 5     2018-11-17  2018-11-17
#> 6     2019-04-28  2019-04-28
#> 7     2016-10-23  2016-10-23
#> 8     2017-08-04  2017-09-03
#> 9     2017-08-04  2017-09-03
#> 10    2018-08-06  2018-08-06
#> 11    2019-09-10  2019-09-10
#> 12    2020-01-01  2020-01-01
#> 13    2019-01-06  2019-01-06
#> 14    2015-05-07  2015-05-07
#> 15    2015-12-19  2015-12-19
#> 16    2016-11-02  2016-11-02
#> 17    2020-08-11  2020-08-11
#> 18    2015-08-30  2015-08-30
#> 19    2016-05-13  2016-05-27
#> 20    2016-05-13  2016-05-27
#> 21    2016-12-25  2016-12-25
#> 22    2019-10-31  2019-10-31
#> 23    2019-10-30  2019-10-30
#> 24    2020-07-12  2020-07-12
#> 25    2016-02-02  2016-02-02
#> 26    2016-01-18  2016-01-18
#> 27    2017-03-16  2017-03-16
#> 28    2018-01-18  2018-01-18
#> 29    2018-05-03  2018-05-03
#> 30    2020-04-09  2020-04-09
#> 31    2016-11-28  2016-12-06
#> 32    2016-11-28  2016-12-06
#> 33    2019-12-23  2019-12-23
#> 34    2015-10-29  2015-10-29
#> 35    2017-08-27  2017-08-27
#> 36    2019-05-01  2019-05-01
#> 37    2015-02-02  2015-02-02
#> 38    2016-12-14  2016-12-14
#> 39    2019-12-18  2019-12-18
#> 40    2020-04-11  2020-04-11
#> 41    2016-11-16  2016-11-16
#> 42    2018-05-17  2018-05-17
#> 43    2015-08-13  2015-08-13
#> 44    2016-07-01  2016-07-01
#> 45    2018-07-20  2018-07-20
#> 46    2016-03-25  2016-03-25
#> 47    2019-09-24  2019-09-24
#> 48    2015-07-16  2015-07-16
#> 49    2015-02-11  2015-02-11
#> 50    2017-04-17  2017-04-17
#> 51    2017-01-23  2017-01-23
#> 52    2019-06-24  2019-06-24
#> 53    2015-02-28  2015-02-28
#> 54    2016-03-16  2016-06-05
#> 55    2016-03-16  2016-06-05
#> 56    2015-11-02  2015-11-02
#> 57    2017-02-07  2017-02-07
#> 58    2019-01-08  2019-01-08
#> 59    2019-05-07  2019-05-07
#> 60    2015-09-04  2015-09-04
#> 61    2016-03-03  2016-03-03
#> 62    2018-11-02  2018-11-02
#> 63    2019-08-24  2019-08-24
#> 64    2016-01-21  2016-01-21
#> 65    2016-01-24  2016-01-24
#> 66    2016-07-12  2016-07-12
#> 67    2015-10-16  2015-10-16
#> 68    2016-10-22  2016-11-04
#> 69    2016-10-22  2016-11-04
#> 70    2017-08-07  2017-08-07
#> 71    2019-09-24  2019-09-24
#> 72    2020-09-16  2020-12-09
#> 73    2020-09-16  2020-12-09
#> 74    2016-08-27  2016-08-27
#> 75    2015-10-04  2015-10-04
#> 76    2019-04-12  2019-04-12
#> 77    2016-06-02  2016-06-02
#> 78    2016-11-17  2016-11-17
#> 79    2017-04-18  2017-04-18
#> 80    2018-10-01  2018-10-01
#> 81    2017-02-18  2017-02-18
#> 82    2017-10-03  2017-12-13
#> 83    2017-10-03  2017-12-13
#> 84    2020-07-03  2020-07-03
#> 85    2015-11-21  2015-11-21
#> 86    2017-06-21  2017-06-21
#> 87    2017-11-08  2017-11-08
#> 88    2016-12-29  2016-12-29
#> 89    2015-04-28  2015-04-28
#> 90    2016-01-10  2016-01-10
#> 91    2017-10-26  2017-11-08
#> 92    2017-10-26  2017-11-08
#> 93    2016-08-08  2016-10-24
#> 94    2016-08-08  2016-10-24
#> 95    2016-08-08  2016-10-24
#> 96    2020-07-14  2020-07-14
#> 97    2020-10-23  2020-10-23
#> 98    2016-08-06  2016-08-06
#> 99    2019-08-23  2019-08-23
#> 100   2015-07-12  2015-07-12