Compute duration between two dates

This function is meant to be for data frame input only and used with dplyr::mutate() to compute age or duration between two character or Date columns. If a vector of breaks is given, the output will be converted to factor with labels generated automatically.

Usage

compute_duration(
  from,
  to,
  lower_brks = NULL,
  unit = c("year", "day", "week", "month"),
  trans = FALSE,
  .transfn = lubridate::ymd,
  verbose = getOption("odcfun.verbose"),
  ...
)

Arguments

from

A character or Date vector for start dates.

to

A character or Date vector for end dates.

lower_brks

A numeric vector for lower breaks passing to the base base::cut() function to convert the numeric result to a factor. The level labels will be auto generated. For example, the level labels are c("<19", "19-24", "25-34", "35-44", "45-54", "55+") for lower_brks = c(0, 19, 25, 35, 45, 55). Default is NULL (no conversion).

unit

A character string specifying the unit of the output. One of "year" (default), "day", "week", or "month".

trans

A logical for whether transform both from and to with the .transfn function

.transfn

A function for transforming the inputs. Default is lubridate::ymd().

verbose

A logical for whether print summary of the out and warning for missing values. Default is fetching from options. Use options(odcfun.verbose = FALSE) to suppress once and for all.

...

Additional arguments passing to base::cut().

Value

A numeric or factor vector of the duration.

Examples

# toy data
n <- 5
df <- data.frame(id = 1:n,
start_dt = sample(seq(as.Date("1970-01-01"), as.Date("2000-12-31"), by = 1), size = n),
end_dt = sample(seq(as.Date("2001-01-01"), as.Date("2023-12-31"), by = 1), size = n))

# get age group at a cut-off
df %>% dplyr::mutate(
 age_grp = compute_duration(start_dt, "2023-01-01", lower_brks = c(0, 19, 25, 35, 45, 55))
)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   23.55   26.92   32.80   36.64   49.18   50.74 
#>   <19 19-24 25-34 35-44 45-54   55+ 
#>     0     1     2     0     2     0 
#>   id   start_dt     end_dt age_grp
#> 1  1 1990-03-16 2012-09-28   25-34
#> 2  2 1999-06-13 2011-01-18   19-24
#> 3  3 1996-02-01 2017-10-03   25-34
#> 4  4 1972-04-05 2012-06-05   45-54
#> 5  5 1973-10-27 2004-05-09   45-54

# compute gaps between two dates in weeks
df %>% dplyr::mutate(
 gap_wks = compute_duration(start_dt, end_dt, unit = "week")
)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   605.3  1130.7  1176.0  1320.2  1593.1  2095.9 
#>   id   start_dt     end_dt   gap_wks
#> 1  1 1990-03-16 2012-09-28 1176.0000
#> 2  2 1999-06-13 2011-01-18  605.2857
#> 3  3 1996-02-01 2017-10-03 1130.7143
#> 4  4 1972-04-05 2012-06-05 2095.8571
#> 5  5 1973-10-27 2004-05-09 1593.1429