Make test data

Make a toy data set for testing and demo. This is for internal use purpose and not intended to be called by users.

Usage

make_test_dat(
  vals_kept = c("304", "305", 3040:3049, 3050:3059),
  noise_val = "999",
  IDs = 1:50,
  date_range = seq(as.Date("2015-01-01"), as.Date("2020-12-31"), by = 1),
  nrows = 100,
  n_any = 50,
  n_all = 10,
  seed = 2023,
  answer_id = NULL,
  type = c("data.frame", "database")
)

Arguments

vals_kept

A vector of values that suppose to be identified.

noise_val

A vector of values that are not meant to be identified.

IDs

A vector of client IDs.

date_range

A vector of all possible dates in the data.

nrows

Number of rows of the output.

n_any

Number of rows to be identified if the criteria is that if any target column contains certain values.

n_all

Number of rows to be identified if the criteria is that if all target columns contain certain values.

seed

Seed for random number generation.

answer_id

Column name for the indicator of how the row should be identified: any, all, and noise.

type

Output type, "data.frame" or "database".

Value

A data.frame or remote table from 'dbplyr'.

Examples

make_test_dat()
#>     uid clnt_id      dates diagx diagx_1 diagx_2
#> 1    20       1 2018-09-14  3042    3057     999
#> 2     6       3 2016-05-13  3056    3043     999
#> 3    95       3 2016-08-29   999     999     999
#> 4    33       4 2018-01-21  3049    3056    <NA>
#> 5    60       4 2018-11-17   999    <NA>    <NA>
#> 6    51       4 2019-04-28   999    <NA>    <NA>
#> 7    30       5 2016-10-23  3044    3056     999
#> 8    87       5 2017-08-04   999     999     999
#> 9    99       5 2017-09-03   999    <NA>    <NA>
#> 10   61       5 2018-08-06   999    <NA>    <NA>
#> 11   92       5 2019-09-10   999     999    <NA>
#> 12    7       5 2020-01-01  3049    3050     999
#> 13   15       6 2019-01-06   305    3041    <NA>
#> 14   46       7 2015-05-07  3046    3049    3044
#> 15   16       7 2015-12-19  3040    3046    <NA>
#> 16    4       7 2016-11-02  3046    3056    <NA>
#> 17   65       9 2020-08-11   999    <NA>     999
#> 18   98      10 2015-08-30   999    <NA>    <NA>
#> 19   10      10 2016-05-13   304    3050    <NA>
#> 20   13      10 2016-05-27  3042    3040    <NA>
#> 21   72      10 2016-12-25   999     999    <NA>
#> 22   62      12 2019-10-31   999     999    <NA>
#> 23   83      13 2019-10-30   999     999     999
#> 24   97      13 2020-07-12   999     999    <NA>
#> 25   93      14 2016-02-02   999    <NA>     999
#> 26   29      15 2016-01-18  3044    3046    <NA>
#> 27   48      16 2017-03-16  3047    3053    3042
#> 28   37      16 2018-01-18  3041    3054     999
#> 29   64      17 2018-05-03   999    <NA>    <NA>
#> 30   12      17 2020-04-09  3054    3048     999
#> 31  100      19 2016-11-28   999    <NA>     999
#> 32   53      19 2016-12-06   999     999    <NA>
#> 33   39      20 2019-12-23  3043    3057    <NA>
#> 34   71      22 2015-10-29   999     999    <NA>
#> 35   77      22 2017-08-27   999     999    <NA>
#> 36   57      22 2019-05-01   999     999    <NA>
#> 37   59      23 2015-02-02   999     999    <NA>
#> 38   96      23 2016-12-14   999     999    <NA>
#> 39   14      23 2019-12-18  3045    3053     999
#> 40   21      23 2020-04-11  3046    3054    <NA>
#> 41   25      24 2016-11-16  3043    3045    <NA>
#> 42   68      24 2018-05-17   999    <NA>     999
#> 43   27      25 2015-08-13  3046    3058     999
#> 44   94      26 2016-07-01   999    <NA>    <NA>
#> 45   90      26 2018-07-20   999    <NA>     999
#> 46   50      27 2016-03-25  3041    3047    3040
#> 47   43      27 2019-09-24  3054    3044     305
#> 48    5      28 2015-07-16  3045     999    <NA>
#> 49    8      29 2015-02-11   305    <NA>     999
#> 50   22      30 2017-04-17  3051    3048     999
#> 51    1      31 2017-01-23  3058     304    <NA>
#> 52   91      31 2019-06-24   999    <NA>     999
#> 53   45      32 2015-02-28  3052    3042    3042
#> 54   31      33 2016-03-16  3054     305     999
#> 55   70      33 2016-06-05   999    <NA>     999
#> 56   54      34 2015-11-02   999     999    <NA>
#> 57   47      34 2017-02-07  3055    3041    3043
#> 58   69      34 2019-01-08   999     999    <NA>
#> 59   28      35 2019-05-07  3055    3052     999
#> 60   55      36 2015-09-04   999     999    <NA>
#> 61   82      36 2016-03-03   999     999    <NA>
#> 62   26      36 2018-11-02  3058    3047    <NA>
#> 63   18      36 2019-08-24  3050    3053     999
#> 64   42      37 2016-01-21  3049    3041    3050
#> 65   56      38 2016-01-24   999     999    <NA>
#> 66    9      38 2016-07-12  3054    3049     999
#> 67   88      39 2015-10-16   999    <NA>     999
#> 68   23      39 2016-10-22   305    3048     999
#> 69   76      39 2016-11-04   999     999    <NA>
#> 70   24      39 2017-08-07  3053    3047    <NA>
#> 71   52      39 2019-09-24   999     999     999
#> 72   89      39 2020-09-16   999    <NA>     999
#> 73   86      39 2020-12-09   999    <NA>    <NA>
#> 74   79      40 2016-08-27   999     999     999
#> 75   67      41 2015-10-04   999    <NA>    <NA>
#> 76   38      41 2019-04-12  3043    3046    <NA>
#> 77   17      42 2016-06-02  3052    3055     999
#> 78   81      42 2016-11-17   999     999    <NA>
#> 79    3      42 2017-04-18  3052    3044    <NA>
#> 80   74      42 2018-10-01   999    <NA>     999
#> 81   19      43 2017-02-18  3041    3051    <NA>
#> 82    2      43 2017-10-03  3053     305    <NA>
#> 83   41      43 2017-12-13  3058    3045    3043
#> 84   85      43 2020-07-03   999    <NA>     999
#> 85   58      44 2015-11-21   999     999     999
#> 86   73      44 2017-06-21   999    <NA>     999
#> 87   35      44 2017-11-08   304    3053     999
#> 88   40      45 2016-12-29  3055     999    <NA>
#> 89   63      46 2015-04-28   999     999    <NA>
#> 90   11      46 2016-01-10  3054    3046     999
#> 91   75      46 2017-10-26   999     999    <NA>
#> 92   49      46 2017-11-08  3055    3059    3051
#> 93   80      47 2016-08-08   999    <NA>    <NA>
#> 94   36      47 2016-08-21  3043    3058    <NA>
#> 95   78      47 2016-10-24   999     999     999
#> 96   66      47 2020-07-14   999     999    <NA>
#> 97   44      47 2020-10-23  3043    3054    3040
#> 98   34      48 2016-08-06  3053    3056    <NA>
#> 99   32      49 2019-08-23  3058    3049    <NA>
#> 100  84      50 2015-07-12   999     999     999